satellite/jobq: get rid of funnel goroutine
Instead of a watcher goroutine that moves jobs from the retry queue to
the repair queue when they become eligible, we can just move all such
jobs before doing any operation where it might matter. This
significantly simplifies the lock and goroutine-communication scenario,
and ought to resolve a deadlock we are seeing in production with a full
headChan channel.

This may increase general operation latency slightly, but total amount
of work done by the cpu should remain about the same.

Change-Id: I9fe49661553e46be5365d37e134d6cd519f54812