storagenode/db: make the db_stat metric opt-in
We have duplicated db_stat metrics due to the migration (which prefers to open/close databases), and Alloy/Prometheus doesn't really like it...

We register db_stats metrics during DB open, but we couldn't easily remove it from monkit.

We already introduced a function to open database without metrics, but the default value was true (with stat).

This patch reverse the logic: by default no stat, only if it explicitly requested.

This supposed to exclude the migration stats in a more reliable way.

Change-Id: I42db0dd04124e53d012ea3b7260ed1ca996c01bf