shared/modular/cli: helpers to use clingy and mud together
clingy is a better fit for mud as it can handle dynamic configuration better, compared to cobra.

With this approach, it's easy to regsiter mud componens as subcommands (RegisterSubcommand)

Examples can be found in mudrunner.go.

One additional benefit of this approach: it's possible to defined the required modules with CLI arguments (not only with environment variables)

Change-Id: I5cef957dc170d1fc4c8d591febe3374a1cfd8b47