{cmd|satellite}: Split get segment for Audit & Repair
Create 2 new methods to get a segment by position from metabase.

One method is used for Audit and the other for Repair. Each method
returns its own type, currently Audit is a type based on the Repair one.

The new return types don't have the encrypted fields that the `Segment`
type returned by the generic get segment by position method.

The goal of having 2 separated methods and return types for Audit and
Repair is to be flexible in the future to trim fields on each one
independently for optimization purpose.

This commit only brings the changes to replace the calls of the original
method by these ones and adapt the related functions and methods that
uses the returned segment. The DB access is kept and these to new
methods use the former method and convert the result to the new types.

Another commit will contain the changes in the database adapters to
optimize the queries for each one only retrieving the needed columns.

Related to https://github.com/storj/storj/issues/7491

Change-Id: Ibc14ec2e2d97623adbf21ded028b6ffe42d89247