storagenode/hashstore: optionally fsync every write
this can prevent data loss on servers that don't
have UPS or whatever and have unclean shutdowns.

this change also cleans up the log writer a bit
now that every code path goes through the in
memory buffer, as well as removing the extra code
around the hashtbl being full as that doesn't
really happen anymore and adding it in the first
place was a bit of an overreaction to a bug.

Change-Id: I58f7bae59abc7cf2bdcc6316a8a19b0a76866e41