dev: use local-first lint configuration (with go tool)
Today we store the Linter configuration in storj/ci repository.

This makes it less visible, and less usable for all the IDE tools and for AI agent tools.

Importing it to the repo itself makes it more explicit, still we can bump rules with downloading the latest version from ci repository.

Also: it makes it possible to use the new `go tool` which makes linting faster (doesn't require to download huge Docker container)

Change-Id: Ib63c2b51486cc9391442dc422ef2e2a8396d376b