storagenode/hashstore: fix potential log corruption from rewrite
when we rewrite a value from a log during
compaction, we change the seek position of the log
file we're reading from in order to get
copy_file_range optimizations. this is normally
fine because 1. this only happens during
compaction and so no one else could be writing
from the log file and 2. all readers use the
ReadAt methods that ignore the seek position.
however, if compaction is canceled or errors in
some way, then it's possible that the log file is
replaced with a seek position that isn't the end
and so the next write to it will overwrite some
existing data.

the fix in this commit is to always seek to the
end of the log file before writing to it and using
that as the offset. this way there's no way it can
be wrong. we then reclaim the space during writer
errors on close by using the truncate call and
relying on the next write performing the seek to
the end to update the pos. it's ok if truncate
errors because log files are robust to
unreferenced garbage data in them and since we
always defensively seek, the pos is allowed to be
arbitrary.

Change-Id: I1f88493d5c03e85f45781436d2b97688127b9152