satellite/orders: Use gomock maintained fork
Use the gomock maintained fork, the previous one is deprecated in favor
of the Uber which is currently maintained.

Anyway without using the new fork, the `go:generate` directive failed
for me without adding the new parameters, so this directive had to be
updated despite of using the unmaintained version.

Change-Id: I379ca4639c76187c1f3b9e83907d55bbab90a62a