satellite/repair: Pass nodes getter to repair orders
Change the orders for repair function to receive a function as a
parameter to get the nodes to allow the callers to chose how that getter
behaves.

Satellite repairer uses this function and for now it passes the same
overlay method that the orders service method was using.

In upcoming commits repairer will replace this function to reduce the
number of database hits.

This commit involves changes in the `orders` package, but `repair`
changes are what produced the `orders` changes.

Change-Id: I82a0aa61324c2d0077cac857761fcd2b2f4ebe64