satellite/repair/repairer: Cache nodes for orders
Retrieve all the online nodes and cache them in memory so repairer
doesn't need to get the nodes from the database for each segment to
repair and reduce hits and bandwidth from the database.

Add configuration fields to configure when the cache needs to be
refreshed.

This commit implies to add a new query to retrieve all the online online
nodes for repair because the existing method only retrieve the ones with
the passed IDs.

The `repairer` changes implied changes in other packages.

Change-Id: I4f19c363d8b1690fe57a6c4c66e6ba7d4afa176b