satellite/orders: Split get Audit & Repair nodes
Split the method that returned a list of nodes reputation for Audit and
Repair into two different ones to easily change Audit or Repair without
breaking one when changing the other.

We want to optimize Repair. This split allows to make changes without
having to thing in the compromise of fulfilling the needs for both and
be able to apply specific optimizations.

We can take a look after the optimization if we can unified again or if
we can optimize Audit.

The method that retrieved both is renamed with "Test" prefix because
after it is used to check data on tests and for now we want to leave
those checks as they are. We will look at remove this method and
refactor the tests checks to use new methods if it's possible.

This commit implies changes in several components, but the `orders`
package is what triggers the changes in the other packages.

Change-Id: Ia5726829a1ebb4f0e4c6bdf158ddd8cbe4cc89fd