satellite/orders: Remove test method
Remove a test method that was created when we split the method to get
order for Audit and Repairs to avoid changes in existing tests.

After the split, we could verify that that this method could be replaced
by the method to get order for repair in those tests, so it isn't needed
for anything.

This commit involves changes in several packages because deleting the
mentioned method implied the deletion of its counterparts in other
packages and mocks.

Change-Id: I4b8a27aa09d67e28dc1c33545d957c5f2900f89b