satellite/metainfo: fix flakiness of TestEndpoint_CopyObject
This change fixes an issue that caused inconsistent failures for the
TestEndpoint_CopyObject unit test.

The test retrieved a list of all objects from the database in several
places and expected this list to be the same in each execution of the
test. However, this was not the case: objects were sorted according
to their stream IDs, which contained random values. This caused
portions of the test code to reference incorrect objects and
subsequently fail.

Change-Id: I5f0acd329ae3563ee08589fc0f80d5ea214a1338