satellite/metainfo: fix flakiness of TestEndpoint_checkRate
This change fixes an issue that caused inconsistent failures for the
TestEndpoint_checkRate unit test.

This test verifies that the metainfo request rate limiter works
correctly by making multiple requests to the metainfo API in quick
succession, expecting them to be handled close enough to each other in
time that the rate limiter is triggered. However, in slow test
executions, the requests may be handled so far apart that the rate
limiter isn't triggered, causing the test to fail.

Change-Id: I8cf1a9a8c7a562cc964f641385616f48f147e92b