satellite/metainfo: fix flakiness of TestEndpoint_Object_Limit
This change fixes an issue that caused inconsistent failures for the
TestEndpoint_Object_Limit unit test.

This test verifies that the object upload rate limiter works
correctly by issuing multiple object upload requests to the metainfo
API in quick succession. The requests are expected to be handled close
enough to each other in time that the rate limiter is triggered.
However, in slow test executions, the requests may be handled so far
apart that the rate limiter isn't triggered, causing the test to fail.

Change-Id: Ic0d5156f91cf363c13429a95bbc1071badda1cba