satellite/satellitedb: Unify methods with same logic
Unify two methods that were split to untangle audit and repair to
facilitate a performance improvement in repair.

The method are kept separated at the service layer which is a wrapper of
the database layer to avoid breaking changes of the service consumer in
case that we have to make changes in audit or repair in the future.

The database layer method are unified because the queries are the same
after we verified that the consumers of the audit data doesn't require
the extra columns that the query requested unlike the repair query.

Keeping the 2 methods of the database layer under a unique method and
renaming it accordingly reduce code duplication and cause less confusion
to readers to differentiate between what currently audit and repair
need.

Change-Id: I704aedad8e7768908cd2747d06f230d0879de205