satellite/repair/repairer: Compare segments pieces in DB
The `repairer.Repair` uses an unexported method to compare if a segment
was altered between it was selected for audit.

This method gets all the pieces of the segment and compare them in the
application layer.

We want to reduce Spanner paid bandwidth (egress) to reduce costs, so we
want to make this comparison in the database and avoid the bandwidth of
transferring the pieces from the database to the application.

Spanner bandwidth (egress) has a cost when the data is transferred
outside of Google Cloud. Because we run repairs outside of Google Cloud,
this change will reduce the cost of Spanner paid bandwidth because
ingress is never paid.

This commit adds a new method in metabase to do this comparison in the
database and replace the change the unexported method to use this new
metabase method.

Issue: https://github.com/storj/storj/issues/7522
Change-Id: Ic6cc76e75e86cb62ee8a3103aee5fa6702e3b654