satellite/metainfo: fix more flakiness of TestEndpoint_CopyObject
This change fixes an issue that caused inconsistent failures for the
TestEndpoint_CopyObject unit test. It is a continuation of the work
started in a3f58bac.

If the test uploaded an object during a run of the tally loop, the
storage usage in the live accounting cache would be inaccurate, causing
the test to unexpectedly receive an error indicating that the storage
limit was reached.

Change-Id: I70e9c90abc002c20a0a546520563e02df8989801