satellite/repair/repairer: Use new String method redundancy scheme
RedundancyScheme type got a new method (i.e. `String() string`) to
satisfy the `fmt.Stringer` interface, so we don't need the wrapper type
to provide the capability to it.

This commit bumps common package and removes the wrapper type and the
helper Zap function simplifying and making more clear the
implementation.

Change-Id: Ib40f7d09a2f46e542f437e4537d3f3f05a2ed9bb