private/migrate: check before creating version table
The `CREATE TABLE IF NOT EXISTS` query can be slow in Spanner. Executing this for every migration step introduces unnecessary latency, especially when the table already exists. This change optimizes the process by first checking the schema information to see if the version table is present before attempting to create it. This avoids the expensive query in the common case.

If we will prove that is causing slow migrations then we can think about reorganizing code to make things more optimal.

Co-authored-by: ethan@storj.io

Change-Id: Ic8fc26289dcb13d982f2ab62744de8b9698db6a7