web/satellite: reset card setup on failure
This change will reset the add card form and get a new card setup
secret if an error occurs on our side after card is already added on
stripe. This fixes an issue where the setup intent would have completed
already after an error occurs on the satellite so the user cannot add
the card anymore because the setup intent is already completed.

Change-Id: I6b2abd346032fe607fbaecd31d63061e13d94b90