cmd: Bump Debian version in Docker images
Our build is failing because the `apt-get` command executed inside of
the image is failing.

Debian Buster release is quite old an its "end of life long term
support" ended on 2024-06-30.

It isn't worth it to try to fix the Buster image doing some hacks (if
any) and we should use the last stable release, Bookworm.

Debian releases: https://www.debian.org/releases/

Change-Id: I9041e9127618711248ea39dbcc17045fd7eed28f