web/satellite: fix bulk object deletion within folder
This change fixes an issue in the satellite UI that caused bulk file
deletion to exhibit unexpected behavior when performed inside of
folders.

Only the names of the selected objects were included in the deletion
request rather than their full keys. This resulted in the accidental
deletion of any objects with matching names within the root of the
bucket.

Change-Id: Id77a92c2214ca181beaef740c02346ecc4e551d5