private/testplanet: set default config values as in production
Set storage node config flags in the same way as in production with some overrides that are needed for testplanet. Earlier lots of flags were unset (e.g. 0) and it was causing some side effects e.g. warnings/errors in testplanet logs.

Flags that have the same default value in config as in testplanet
configuration were removed.

Change-Id: Ief1ffd4c599a72beb0b8089c2855c27156e32101