storagenode/storagenodedb: fix CompleteGracefulExit
Now it's possible to disable WAL with the additional txlock immediate
and set max open conns 1.

txlock immediate ensures that when the tx starts in read mode then
it isn't in the middle upgraded to a write tx, which may cause
"database is busy" errors.

The same goes for max open conns, which tries to ensure that we
don't have concurrent access.

Both of these seem to be more prevalent when WAL is disabled.

Change-Id: Ibef69a1e7f98ccfa702805148fd1a0c979a6c9a9