cmd/satellite: add flag to skip db version checks
Rarely we may need this flag to be able to deploy pods with different
image version when there is migration that bumps DB schema version
between versions.

This may be useful during emergencies where we need to do some non
standard steps. Flag needs to be used with caution and with proper
understanding.

Change-Id: Ica0b78694f35c1f0b25909b809bf34c0b536c44f