satellite/emission: fix carbon accounting
the emission package logic assumes a fixed amount of data
is stored for a certain duration, but invoices have the units
MB*months, instead of just MB stored for a month. this is a
problem because x MB stored for a month uses less hard drives
than 30*x MB stored for a day, even though they get billed
for the same usage. put another way, if you fill one 20TB
hard drive and store it all month long, that's 20TB*months
but one hard drive. If you fill 30 20TB hard drives and
store it for a day, that's still 20TB*months, but it took
more hard drives. more hard drives increases the carbon
accounting.

so, the past fix was to try and have the emission package
handle TB*time-units, but the way it was implemented didn't
account for amortized embodied carbon correctly.

this change is a new fix. we're going to just give up on
the problem. we are going to assume that 20 TB*months is
stored as 20TB for one month. even though this is wrong
within a month, over the course of a year, it will be
mostly right, each month is estimated differently.
if a customer adds 20*30 TB at the last day of one month,
we'll estimate the carbon incorrectly due to the above
described problem, but the following month will get it
right. so over time the invoice estimates will be more
right than they currently are.

Change-Id: Id4291ab16312225da2a394f7db73e2f5a8b4003a