satellite/metabase: CopyObject add intermediate pending stage
Currently when an error occurs after inserting the segments and
inside the transaction, we don't cleanup the segments.
By not having any objects with the specific stream id it makes
cleaning up the zombie segments difficult.

If we first insert a pending object, then we can avoid
this problem. Inside the transaction we then only
update the status, usually. In rare circumstances there's
a concurrent upload to the same object key, in which case
we need to create a new version number and reinsert for
Spanner.

Change-Id: I6156bc3bd63138b7311d35d41f4f88dded389e21