satellite/metabase: reduce contention from DeletePendingObject
This is quite aggressive optimization for deleting pending objects.
To avoid contention and increased lock time on Spanner side this
change removes as much places that will be locked by tx as possible.

What's changed:
* remove RETURN clause from object deletion completely
* return only basic object metadata that we knew even before deleting
* change segments deletion to use Spanner Read API

For pending object deletion we can return way simpler result than
for regular delete because object returned in this case is not really
used anywhere on libuplink side.

Change-Id: If378838dd3d3d9248f950103ab388d763a30d61b