satellite/db: fix duplicate usage report rows
This fixes an issue where rows in the usage report will be repeated if
for each bucket the user has, there are at least one matching bucket
for another project. This issue happened because the SQL join that
fetches the bucket information only joined on the bucket name only and
not with the project as well. This is likely the cause of the slow
performance of the usage report as well, since it would return a lot of
duplicate rows.

Issue: https://github.com/storj/storj-private/issues/1428

Change-Id: I3f901fb0ca2c269b160b8d4a6d738f1b9326e05f