cmd/satellite: satellite command to list duplicate CC usages
Added new command which goes through all the customers and prints or generates a CSV of all the CC usages.
There is a required 'minCustomers' argument to filter out CCs that are used by just 1 or 2 customers.
There is a required 'csvPath' argument which could be a path to a CSV file. Additionally, it accepts '-' value to just print the results instead of generating CSV.

Issue: https://github.com/storj/storj-private/issues/1332

Change-Id: I9af2cd41c521938cdb917a6ba890d7355d5900cd