satellite/metainfo: add option to upload with BeginObjectExactVersion
Idea behind this is to randomly generate negative version for pending
objects. Because of that we are removing one subquery from object insertion query. That should improve overall performance and reduce
risk of locks between upload transactions.

Because we are choosing version randomly we my have collisions. To
handle that we are retrying begin object tx if such collision
happens (max 5 times).

New feature is behind feature flag. It can be enabled globally or for
specific project.

Change-Id: I41fa5ed963b6d20d6c4e7404a451af314b84f320