satellite/audit: use real reporter/containment with modular auditor
The main challenge was to handle DBReporter/Containment. We need to disable them for ad-hoc audit (CLI command), but they are required for real audit.

Mud can support multiple implementation for the same interface.
It requires the registration of two implementations (mud.Provide), and the selection of default implementation (mud.RegisterInterfaceImplementation). After this, commands can choose the other implementation.

The challenge here is that we don't have the second implementation, as `db.Containment()` returns with an interface instead of a dedicated implementation. We couldn't differentiate the "real" implementation based on type, as we don't have type name for that one. Unless we create a wrapper type.

That's what happened in this patch: wrapper for real implementation + noop implementation for the ad-hoc cli commands (if required).

Change-Id: I9c82eb56db0c87c9d2f99f282b35fa49e141c406