storagenode/hashstore: error if table missing when logs present
this makes the system more robust to
misconfigurations when attempting to move the
hashtbl to a different disk, as well as to
corruption of the disk where the hashtbl file goes
missing.

additionally, set ordered rewrite to true. it
should be at most of a couple of megabytes during
compaction per TB of data stored under our typical
churn rates and ensures that we do linear reads
out of the disk as much as possible.

Change-Id: If0c53c0b53e86ea6a7a962febfb13c9fc01825da