satellite/nodeselection: support configuring cohorts
this change goes hand-in-hand with
 * https://review.dev.storj.tools/c/storj/common/+/18141
 * https://review.dev.storj.tools/c/storj/uplink/+/18413

with these changes, the uplink will now support flexible
requirements about how many nodes with certain tags must
be uploaded to.

the configuration language has three primitives:
  * minimum node counts
  * ands
  * "withhold"

Here is an example configuration:

    min(49) &&
      withhold(attr("tag:$SIGNER_ZERO/dc"), 1,
          withhold(attr("tag:$SIGNER_ZERO/dc") + "-" +
                   attr("tag:$SIGNER_ZERO/rack"),
                   3, min(29)))

This configuration says that no matter what, there must
be at least 49 successful uploads (first "and" branch).

Separately, it must also be true that, if you withhold the
largest DC, and then after that withhold the largest 3 racks of
the remaining DCs, there are still 29 successful uploads.

Change-Id: I1f3c62bf292bec725129ceb7970bebebd273586e