satellite/metainfo: add support for bursting to single path upload limiting
occasionally, some customer may want to double upload to the same
path. maybe it's a bug on their side, but it's not really our
place to reject it. our problem is if someone does it repeatedly
over a long period of time.

so this change makes it so that uploads to the same path allow for
bursting beyond the limit.

it does it by replacing our basic LRU-based path limiter with our
more fancy token-bucket-based bloom rate limiter.

Change-Id: Ic55cf887bff04bfa8cd6e49961e6bca45b26cf45