cmd/satellite: Homogenize log field user-accounts cmd
The sub-commands under the `user-accounts` command log the user's email
with two different keys, `email` and `user_email`.

Having to different keys for the same value depending of the log message
makes harder to extract data from them easily.

This commit address all the logs with key `user_email` fields to use
`email` as key, making all of them to use the same key for the same
value.

Change-Id: I39effd582a6d7c8dc0ec92532a22a8150414d5e2