satellite/{console,db}: update status_updated_at on project status update
Also add a new ProjectPendingDeletion (2) project status.

This change also fixes an issue in the TestConsoleTx test where the it
will fail because the sync2.Once that is used to initialize the
projects DB object is used inside a transaction first. This causes the
projects DB object to be uninitialised outside of the transaction, and
since the Once is already used, it will not be initialised again.

Related: https://review.dev.storj.tools/c/storj/storj/+/18451/1

Change-Id: I48c5d00749e9a3ecc6f8ac8cc9866e6e13f742a4