changestream: skeleton service for procesing Spanner changefeed
This is the initial skeleton of the service, which can read changestream from a Spanner, and log changes to stdout.

There are multiple missing points (See TODO), and changestream.Processor likely needs a better implementation.

The main focus here was to create the wiring and basic subcommand infrastructure.

To test:
   1. create a changefeed (For example: 'CREATE CHANGE STREAM `bucket_eventing` FOR objects (status, total_plain_size) OPTIONS ( exclude_ttl_deletes = TRUE, exclude_update = TRUE);'')
   2. Create a config file (and identity). It's enough to include metabase.database-url as the config value.
   3. Start service with modular satellite: `go run ./satellite/satellite change-stream --identity-dir ... --config-dir ...`

https://github.com/storj/roadmap-private/issues/464

Change-Id: I02d1c49076a8cb268d0efd9893bb7e89889f2c1c