mudplanet: helpers to make e2e tests with mudplanet
BACKGROUND:

mudplanet is like the testplanet, but it supposed to be more lightweight and flexible.

One main goal is to keep ot fully generic: We couldn't use go test cache very effectively, as testplanet depends on the whole world (=planet).

If something is changed in the storagenode side, ALL tesplanet tests (and the related packages) should be re-executed.

mudplanet is different. It can execute any components. Therefore it doesn't have dependencies.

It supports partial services, like starting a Satellite without DB or just with the required services.

THIS PATCH:

This patch adds new helpers, to make the testing easier.

The main improvements is the `TestUploadInline` in smoke_test.go. All other go is just making it possible.

 1. Muliple helpers and structs in planet.go
 2. planet.go supports run wrapper (seems to be a big change, but just identification)
 3. mudplanet/satellitetest: helper to run test with multiple different databases.

Change-Id: I0928d56760ca71c7c28ba32ea9688bc14b0fbc65