go.mod: Indicate to use Go 1.24.7
Go was set to 1.24.2 and it was reported by Docker Scout that we had to
use at least go 1.24.4 standard library to fix CVE-2025-22874.

Bump Go to the last patch release on the same minor because our build
it's still using Go 1.24.

This commit set the same version to other go.mod files present in this
repository.

Change-Id: I04b6839b948e710890cb9784660b2a40d2b4acd4