apigen: fix gen api bugs
This fixes a bug where endpoints in the form /path/{param}/extra will
be generated as /path/{param}.

It also fixes a bug where an endpoint parameter with a custom type
declared in the API's PackagePath will not be generated correctly.
E.g. type *pkg.CustomType will be generated as *pkg.CustomType instead
of just *CustomType.

It also exports the generated API error class.

Change-Id: I04f18dd2a94ea228cb060d2b67d9768f59030d58