storagenode/hashstore: reduce open file handles
this makes it so that when a log file is over the
max log size, we close it. for reads, we always
open a new file handle through a small lru of
default size 10. this should reduce memory usage
on windows which uses a lot of memory per file
handle for some reason, while still having good
performance everywhere.

Change-Id: I746b82cc7ef9bfa84ca5219b043d683b196e0c1f