satellite: set project default placement
This change updates project creation in admin and console to set default
placement to the first of the configured allowed placements entitlement
on project creation if entitlements are enabled.
Metainfo is also updated to not require projects to have default
placement 0 to be able to use self-serve placement if entitlements are
enabled.
Finally, the UI is updated to not require default placement 0 for new
projects to use self-serve placement if entitlements are enabled.

Issue: https://github.com/storj/storj-private/issues/1549

Change-Id: Iad5c4610d929d9e6d1938e3c4dcfdaf1474b6dc5