satellite/payments: support unit conversion at the time of invoice generation
Added support for defining a product where price and usage on the invoice is in terms of GB (egress) and GB-month (storage) instead of MB and MB-month.
Added logic to always round up the egress/storage usages.
So, if there is at least a 1 byte of usage it's going to be rounded up to 1 GB

Issue: https://github.com/storj/storj-private/issues/1551

Change-Id: I8ca2b03f45f14643da37498d3f6cd191faf37933