satellite/repair/repairer: fix memory usage for in memory download
This part of code was recently changed but it was bad choice.
Butes.Buffer was using lot of additional memory because it was trying
to grow it's internal buffer. Previous code was working better. One
improvement that was added here on the top of old code is buffer
allocation in advance. It will be more optimal to read directly to
full buffer instead growing it on demand with io.ReadAll because we
know piece size from the beginning.

Change-Id: I54baf45ac636e0185e38ea4561349c4f0f22daea