satellite/nodeselection: support invariant filters
For a placement, invariant should be true. If not true, the pieces will be marked as unhealthy but retrievable (clumped).

This can be used to initiate moving out pieces from selected nodes.

Note: this is different from filtering, as filter applies to downloads as well, which may make the segments unavailable.

Change-Id: Idf7a115de3eadd59dd717c68ea980324a5b69022