satellite: Untangle Overlay Service from Repair Checker
Untangle Overlay Service from the Repair Checker to see what methods it
depends considering that the Overlay Service has many.

Add a new configuration parameter to set the nodes online window
specifically for repairs to be able to distinguish the online window
between the repair processes and others.

Add new method to Overlay as on homologous of an existing one, but the
new one uses the new online window for repairs.

This new method is used by Repair Checker, as it was using the
homologous existing one because it needs to use the new online window
for repairs configuration parameter.

Change-Id: I83451413c8f31a96d8d3e8326b7feed9bcb2a719