satellite: Untangle Overlay Service from Repair Repairer
Untangle Overlay Service from the Repair Repairer to see what methods it
depends considering that the Overlay Service has many.

Add new methods to Overlay as on homologous of an existing ones, but the
new ones use the new online window for repairs.

This new method is used by Repair Repairer, as it was using the
homologous existing ones because it needs to use the new online window
for repairs configuration parameter.

Change-Id: I6fefe79aeb8affc36231e44fca475b3c23ad1004