web/satellite: update upgrade dialog for new pricing
This change updates the account type and account activation steps of the
onboarding flow to reflect the pricing updates. It mainly achieves this
by unifying the upgrade account flow and account setup steps to use the
same components for selecting a pricing plan and adding payment methods.
The result of this is that the upgrade dialog also now merges the
package plan selection step into info step of the upgrade flow, rather
than being a separate step.

Issue: https://github.com/storj/storj-private/issues/1568

Change-Id: I468865f9a192452f5db5d28bc90d935b7fc9a583