satellite/metabase: fix version handling returned by PrecommitConstraint
With negative version for pending objects it was possible to get
HighestVersion from PrecommitConstraint. Not all cases were adjusted
to new logic and it was possible to have committed object with
negative version.

This change makes sure that PrecommitConstraint never returns
negative HighestVersion which makes later easier to determine final
version for committed object.

Change also enabled TestingAlternativeBeginObject as it's
configuration that is used in production right now.

Change-Id: I195f0229adf9d609a87d286b34e1366efc6cde43