satellite/{console,payments}: fix entitlements for users with non-zero default placement
This change fixes a bug where projects created by users with non-zero
default placement will be entitled to self serve placement selection and
have a default placement other than the user's default placement.
This change makes it so that in such a case, the project's default
is set to the user's default placement. And the entitlements list of
allowed placement is set to a list of only the user's default placement.

Issue: https://github.com/storj/storj-private/issues/1579

Change-Id: Ie0ef377e43ea2c7a777df12463277da772f7651c