satellite/audit: convert reporter_test to use mudplanet instead of testplanet
As I would like to run modular based auditor everywhere, it's better to run the mud version of auditor.

As a result it's 2x times faster (example: TestReportingAuditFailureResultsInRemovalOfPiece is 4s instead of 9.5) as it doesn't require storagenodes, no need to wait for the first sync. Also: it doesn't start any irrelevant components, just the auditor (and dependencies).

Change-Id: I258ba4ea715ef44b2a05718605c0d4562679c65f