cmd/satellite: Add a few debug messages to user-accounts
Add a few debug message to the users-accounts commands to reflect
progress.

Currently the commands are silent if there aren't errors, which causes a
sense that the process is stuck, or perhaps it's, but we don't have any
easy sign.

This commit also replace the String zap logger fields by Stringer for
those that were calling the String method. This helps not to have those
calls in case that there is not login because the set level is higher,
reducing unnecessary overhead.

Change-Id: I65f1a1dc8f05b9ce3f3a1f8dc69cf532784c6956