satellite/reputation: force update vetted_at when required
We have vetted_at status both in `reputations` and `nodes` table.

After reputation is updated, the code checks if the nodes table should also be updated.

But changes for the vetted_at field is accepted only, if it's done by the same goroutine (or updated at the same second).

This is fragile. In case of the update of the `nodes` table is failed, it won't be fixed any more, as all subsequent changes will have an older `vetted_at` changes.

While the root cause how it can happen not fully clear (it can happen during auditor shutdown, but we have way more problematic nodes. Likely related to higher number of concurrent auditor workers...), it would be more safe to always update nodes table when vetted_at is not yet set, but it should be.

Change-Id: I0e30124ccbf33449d77745832980764eb744de67