storagenode/hashstore: write hint files for fast fsck
writes out a hint file in the meta directory that
contains which log files are mutable at some safe
point in time. this will allow a future fsck to
skip over immutable log files when checking the
integrity of the log files with the hashtbls.

most of the change is a refactoring of the code
used to filter through and parse file names as the
hint file was doing the same thing as the log and
hashtbl files, so the shared code became apparent.
this is now in its own file with tests.

Change-Id: I55b05fefa638191973b1e57d7c647e31e06cae7c