satellite/{stripe,web}: add report download to billing history
This change allows the user to download the usage report for the period
of an invoice.
For this to work, the period start and end dates are corrected to use
the dates in the invoice item instead of the dates in the invoice itself
because that is the date the invoice was created and not the usage
period.
It also updates the header of the date column of the billing history
table to "Usage Period" to be clear what that invoice is for.

Issue: https://github.com/storj/storj-private/issues/1546

Change-Id: If03e4437a1377c74368e8ae036252b3e15ec6626