satellite/eventing: fix version id in s3 events
So far the S3 events included set the object version id to version_id
column, which is incomplete. The complete object version id is the
StreamVersionID type, which a combination of the stream_id and
version_id columns.

To provide the complete object version id in the S3 events, the change
stream needs also watch the stream_id column.

Change-Id: I460b86e6784dbbe5bef30717aaaaebeac66b4ca4