storagenode/hashstore: introduce RecordTails
when doing a quick check on a log, we'll read the
end to get the last written records and compare
them to the largest records that the current table
has. if they are the same, we consider the data to
be intact. this should be cheap because we only
need to check the log files that were mutable
since the last hint file was written.

Change-Id: I9581477c3d9204bc20bfe4208537c59eea27bea1