satellite/durability: enable placement check and declumping for durability check
We kept them false to avoid false alarm in case of healthy but retrievable pieces.

But since then, we generate two reports: one for healthy and one for retrievable.

It makes more sense now to enable it by default.

(Technically I can make it configurable, but I don't see any reason to turn it off in the future).

Change-Id: Ie58859fd0ec53af4c7e7f29d7193404e13682e37