satellite/admin: handle nullable fields properly
This change improves handling of nullable fields by allowing them to be
set to null with "-1", which now also allows them to be set to "0". It
also correctly returns the raw state of these fields when fetching a
project instead of returning their configured default values when they
are null.

Change-Id: I9856548389391766c5de4d62755953ccc14a54ca