satellite/eventing: determine event name from transaction_tag
This helps to better determine the S3 events to emit for each object
transaction we have.

For example:
- Only explicit object deletes must emit a ObjectRemoved:Delete event.
Overwriting an object is not an explicit delete - it should emit an
ObjectCreated:Put event, but not an ObjectRemoved:Delete event.
- Similarly moving an object should emit only a ObjectCreated:Copy
event, but no ObjectRemoved:Delete

Switching the value capture type to NEW_ROW_AND_OLD_VALUES helps with
seeing more the necessary information in the change stream record for
UPDATE mod types where we only change a single column like `status`.

Test data was added to cover more transactions we have on the
objects table.

Change-Id: Ife3117700e8bed1dfa0f8410529cf8056f6c8aac