storagenode/modular: dedicated subcommand for select and normal run
Earlier we tried multiple approaches to define which components should be run.

  1. We tried to tag all the required services with `Service` mud annotation (similar to peer.Services)
  2. We tried to manually define all the components with a `@hashstore` preset
  3. Finally we adjusted the `@hashstore` preset manually

I am not happy with any of these, and I think it's better to create subcommands, which can define all the required components with typesafe function calls.

This change includes two new subcommands `storagenode run` and `storagenode select` for the modular storagenode binary to run required components.

Change-Id: Ie30fe382d020e494bf2c44a3500b0e73727c1771