script: publish rc tags as pre-release
Draft builds are good for internal testing, but for builds that are advertised as minimum/suggested versions on version.qa.storj.io, they should be publised and accessible for qa storagenodes to download the build with our standard node updating processes. This requires the tags to be published and publicly accessible.

Change-Id: Ib6baf192d0f010863262080e635214f107904c12