satellite: modular definitions for web console
This patch contains module definition for the packages which are required by the web console.

We new subcommand on modular satellite can be used to start both UI and backend of the Satellite web console.

(this is not possible with current Satellite, which keeps the backend in the API service)

The webconsole can be started, but further testing is required of all the components are migrated.

Unfortunatelly a few workarounds were required due to circular dependencies. These can be further simplified after the default peers/api pods were removed.

(It's hard to do now, as it would require a big refactor on the old, monolith side)

Change-Id: Ia1bd7dfa59d82e86c38b8e78dd8fac0921f766ac